Complete ProjectMetadataController
This is the mechanism that configuration cache uses to serialize the dependency management state of a project during
project isolation. The existing serialization scheme was quite incomplete and did not serialize the entire set
of project state data. This commit enhances ProjectMetadataController to serialize all component and variant state,
thus bringing the serialized data to parity with the original data.

We go further and update IP to use this state in a load-after-store manner, so that we ensure this data is
serialized properly -- just like we do with load-after-store elsewhere