Progress creating ABI stub variant
We have a new apiStubElements configuration exposing ABI stubs of each project. This currently leverages
Kotlin's ABI stub generation, and exposes Java projects' non-stub implementation artifacts. A future change
should convert the ShrinkPublicApiClassesTransform to a task so that Java can also expose similar stubbed
artifacts instead of their full implementation.

We have configured the compileClasspath of each project to prefer the stubbed classes if available. We are
running into issues. It seems that when kotlin-dsl-tooling-builders compiles against kotlin-api, it fails to
see some classes from it. Looking at the contents of kotlin-abi/build/generated/kotlin-abi, it appears at first
glance that those clases are missing. Why are those classes not in the kotlin ABI outputs? Is this a bug in
the ABI generator? Or, for some reason are these classes not considered ABI?

Running :public-api:jarGradleApi reproduces the problem we are seeing currently with kotlin compilation